+++ doc/source/history.rst
@@ -8,6 +8,8 @@ Release history
 
 * ADDED: argspecs for snippets and modules
 
+* ADDED: more tests for cluda.functions
+
 * FIXED: incorrect behavior of functions.norm() for non-complex arguments.
 
 * FIXED: undefined variable in functions.exp() template (reported by Tibault North)